Ticket #307: fix issues detected with static analysis tools
(Prio 1 open only)

+++ csrc/crypto/sopc_secret_buffer.c
@@ -50,7 +50,7 @@ SOPC_SecretBuffer* SOPC_SecretBuffer_New(uint32_t len)
         if (NULL != sec)
         {
             sec->len = len;
-            sec->buf = malloc(len);
+            sec->buf = malloc((size_t) len);
             if (NULL == sec->buf)
             {
                 free(sec);
